Change ProcessedBsm to a Feature.  Move metadata properties from ProcessedBsmCollection to BsmProperties
