+++ jpo-geojsonconverter/src/test/java/us/dot/its/jpo/geojsonconverter/converter/bsm/BsmProcessedJsonConverterTest.java
@@ -18,7 +18,7 @@ import com.networknt.schema.ValidationMessage;
 import us.dot.its.jpo.geojsonconverter.partitioner.RsuLogKey;
 import us.dot.its.jpo.geojsonconverter.pojos.geojson.Point;
 import us.dot.its.jpo.geojsonconverter.pojos.geojson.bsm.DeserializedRawBsm;
-import us.dot.its.jpo.geojsonconverter.pojos.geojson.bsm.ProcessedBsm;
+import us.dot.its.jpo.geojsonconverter.pojos.geojson.bsm.ProcessedBsmCollection;
 import us.dot.its.jpo.geojsonconverter.serialization.deserializers.JsonDeserializer;
 import us.dot.its.jpo.geojsonconverter.validator.JsonValidatorResult;
 import us.dot.its.jpo.ode.model.OdeBsmData;
@@ -55,7 +55,7 @@ public class BsmProcessedJsonConverterTest {
         deserializedRawBsm.setOdeBsmData(odeBsmPojo);
         deserializedRawBsm.setValidatorResults(validatorResults);
 
-        KeyValue<RsuLogKey, ProcessedBsm<Point>> processedBsm = bsmProcessedJsonConverter.transform(null, deserializedRawBsm);
+        KeyValue<RsuLogKey, ProcessedBsmCollection<Point>> processedBsm = bsmProcessedJsonConverter.transform(null, deserializedRawBsm);
         assertNotNull(processedBsm.key);
         assertEquals(new RsuLogKey("10.11.81.26", null, "12A7A951"), processedBsm.key);
         assertNotNull(processedBsm.value);
@@ -72,7 +72,7 @@ public class BsmProcessedJsonConverterTest {
         deserializedRawBsm.setOdeBsmData(odeBsmPojo);
         deserializedRawBsm.setValidatorResults(validatorResults);
 
-        KeyValue<RsuLogKey, ProcessedBsm<Point>> processedBsm = bsmProcessedJsonConverter.transform(null, null);
+        KeyValue<RsuLogKey, ProcessedBsmCollection<Point>> processedBsm = bsmProcessedJsonConverter.transform(null, null);
         assertNotNull(processedBsm.key);
         assertEquals(new RsuLogKey(null, null, "ERROR"), processedBsm.key);
         assertNull(processedBsm.value);
@@ -91,7 +91,7 @@ public class BsmProcessedJsonConverterTest {
         deserializedRawBsm.setValidatorResults(validatorResults);
         deserializedRawBsm.setFailedMessage("{");
 
-        KeyValue<RsuLogKey, ProcessedBsm<Point>> processedBsm = bsmProcessedJsonConverter.transform(null, deserializedRawBsm);
+        KeyValue<RsuLogKey, ProcessedBsmCollection<Point>> processedBsm = bsmProcessedJsonConverter.transform(null, deserializedRawBsm);
         assertNotNull(processedBsm.key);
         assertNotNull(processedBsm.value);
         assertEquals("{", processedBsm.value.getValidationMessages().get(0).getMessage());
