+++ jpo-geojsonconverter/src/test/java/us/dot/its/jpo/geojsonconverter/serialization/JsonSerdesTest.java
@@ -7,7 +7,7 @@ import org.junit.Test;
 
 import us.dot.its.jpo.geojsonconverter.pojos.geojson.LineString;
 import us.dot.its.jpo.geojsonconverter.pojos.geojson.Point;
-import us.dot.its.jpo.geojsonconverter.pojos.geojson.bsm.ProcessedBsm;
+import us.dot.its.jpo.geojsonconverter.pojos.geojson.bsm.ProcessedBsmCollection;
 import us.dot.its.jpo.geojsonconverter.pojos.geojson.map.ProcessedMap;
 import us.dot.its.jpo.geojsonconverter.pojos.spat.ProcessedSpat;
 import us.dot.its.jpo.ode.model.OdeBsmData;
@@ -53,7 +53,7 @@ public class JsonSerdesTest {
 
     @Test
     public void testProcessedBsmSerdes() {
-        Serde<ProcessedBsm<Point>> serde = JsonSerdes.ProcessedBsm();
+        Serde<ProcessedBsmCollection<Point>> serde = JsonSerdes.ProcessedBsm();
         assertNotNull(serde);
     }
 }
